requirements.txt
fastapi
uvicorn
openai
pytesseract
pillow
faiss-cpu
sentence-transformers
python-multipart
aiofiles
